Refactoring of data access.
The DataManager now doesn't access the data directly but instead only requests the data from a repository, which then accesses and serves the data.

The DataRepository is an Interface, which can be implemented to enable multiple different sources of data.

For now, a DiskDataRepository has been created, which accesses and serves local disk data. This is basically the code, which has previously been part of the DataManager. There is room for improvement.

In the future, an SQLRepository will be added to access and serve data from an SQL-Database.