Adding MySQL support
Added an SQLRepository, which accesses and serves data from an SQL database.
The database is accessed by an SQLHandler, which establishes and manages the connection to the database.
It also serves as an easy way to make queries, updates and commands.
These however, have to passed as raw SQL-Commands. Creating these commands is the job of the SQLRepository.

The IOCallback is now being used with a simple Exception, instead of just an IOException as different sources of data may raise different kinds of exceptions.

Full support for MySQL is now established.

The next commit will make the plugin load a config.yml to retrieve sql credentials.