MySQL support; Making REM work (#10)
* Refactoring of data access.

The DataManager now doesn't access the data directly but instead only requests the data from a repository, which then accesses and serves the data.

The DataRepository is an Interface, which can be implemented to enable multiple different sources of data.

For now, a DiskDataRepository has been created, which accesses and serves local disk data. This is basically the code, which has previously been part of the DataManager. There is room for improvement.

In the future, an SQLRepository will be added to access and serve data from an SQL-Database.

* Adding MySQL support

Added an SQLRepository, which accesses and serves data from an SQL database.
The database is accessed by an SQLHandler, which establishes and manages the connection to the database.
It also serves as an easy way to make queries, updates and commands.
These however, have to passed as raw SQL-Commands. Creating these commands is the job of the SQLRepository.

The IOCallback is now being used with a simple Exception, instead of just an IOException as different sources of data may raise different kinds of exceptions.

Full support for MySQL is now established.

The next commit will make the plugin load a config.yml to retrieve sql credentials.

* Added a conifg file for MySQL-Credentials

Additionally, added connection messages

* Made the rem command work

Simply made the rem command work as it has been started.

It works both using uuid and name

In the future, refactoring of the onCommand method is severely necessary

* Made mysql optional through a config setting

* Removed wrongly placed comment

* Code cleanup

Removed unnecessary access modifiers

* Fixing LGTM Issue

Closing every result-set (using try-with-resources) to make the statement close to make the lgtm issue disappear

Also removed password information from being logged.